refactor: extract common query logic in custom hook
